Fixed runtime interface code. The second.

+++ runtime/interface.py
@@ -73,7 +73,7 @@ class Solver:
 
         input_packed = input_data.packed()
         assert (len(input_desc) / 8).is_integer()
-        input_bytes = int(input_packed).to_bytes(len(input_desc) / 8, byteorder='little')
+        input_bytes = int(input_packed).to_bytes(int(len(input_desc) / 8), byteorder='little')
         for bi, b in enumerate(input_bytes):
             self._input_buffer[bi] = b
 
@@ -83,7 +83,7 @@ class Solver:
         output_desc = data_desc.get_output_desc(self._config)
         output_raw = BitVector(len(output_desc)).create_instance()
         assert (len(output_desc) / 8).is_integer()
-        output_raw.next = int.from_bytes(self._output_buffer[0:len(output_desc) / 8], byteorder='little')
+        output_raw.next = int.from_bytes(self._output_buffer[0:int(len(output_desc) / 8)], byteorder='little')
         output_raw._update()
 
         output_data = output_desc.create_read_instance(output_raw)
